fixed some type conversion conflicts (reverting some previous edits related to this), hopefully resolving some incompatibilities between new exposure mapping approach for MCR and MEL, that didn't work with Munich --- local maven builds seemed to work, so we'll see if this compiles while I run the local offline exposure runner for Melbourne

+++ extensions/health/src/main/java/de/tum/bgu/msm/health/data/PersonHealth.java
@@ -76,7 +76,7 @@ public interface PersonHealth extends Person {
     void resetHealthData();
 
     //TODO: these are methods currently used for Munich health, need to adapt Munich health model
-    void updateWeeklyPollutionExposures(Map<String, Double> exposureMap);
+    void updateWeeklyPollutionExposures(Map<String, Float> exposureMap);
 
     void setRelativeRisks(Map<String, Float> relativeRisks);
