+++ useCases/munich/src/main/java/de/tum/bgu/msm/scenarios/healthMuc/PersonHealthMuc.java
@@ -289,8 +289,8 @@ public class PersonHealthMuc implements PersonWithSchool, PersonHealth {
     }
 
     @Override
-    public void updateWeeklyPollutionExposures(Map<String, Double> newExposures) {
-        newExposures.forEach((k, v) -> weeklyExposureByPollutant.merge(k, v.floatValue(), Float::sum));
+    public void updateWeeklyPollutionExposures(Map<String, Float> newExposures) {
+        newExposures.forEach((k, v) -> weeklyExposureByPollutant.merge(k, v, Float::sum));
     }
 
     public float getAllCauseRR() {return allCauseRR;}
